separate internal classes with that of client one via package
