+++ src/test/java/software/aws/lambda/logging/internal/LambdaAspectTest.java
@@ -1,4 +1,4 @@
-package software.aws.lambda.logging;
+package software.aws.lambda.logging.internal;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
