+++ src/test/java/software/aws/lambda/logging/internal/PowerLogToolEnabled.java
@@ -1,8 +1,8 @@
-package software.aws.lambda.logging;
+package software.aws.lambda.logging.internal;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
-import software.aws.lambda.logging.client.PowerToolsLogging;
+import software.aws.lambda.logging.PowerToolsLogging;
 
 public class PowerLogToolEnabled implements RequestHandler<Object, Object> {
