+++ src/test/java/software/aws/lambda/logging/internal/PowerLogToolEnabledForStream.java
@@ -1,13 +1,15 @@
-package software.aws.lambda.logging;
+package software.aws.lambda.logging.internal;
 
 import java.io.InputStream;
 import java.io.OutputStream;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestStreamHandler;
+import software.aws.lambda.logging.PowerToolsLogging;
 
-public class PowerLogToolDisabledForStream implements RequestStreamHandler {
+public class PowerLogToolEnabledForStream implements RequestStreamHandler {
 
+    @PowerToolsLogging
     @Override
     public void handleRequest(InputStream input, OutputStream output, Context context) {
