+++ src/test/java/software/aws/lambda/logging/internal/PowerToolLogEventEnabled.java
@@ -1,10 +1,12 @@
-package software.aws.lambda.logging;
+package software.aws.lambda.logging.internal;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
+import software.aws.lambda.logging.PowerToolsLogging;
 
-public class PowerLogToolDisabled implements RequestHandler<Object, Object> {
+public class PowerToolLogEventEnabled implements RequestHandler<Object, Object> {
 
+    @PowerToolsLogging(logEvent = true)
     @Override
     public Object handleRequest(Object input, Context context) {
         return null;
